fix: update appveyor to Visual Studio 2017
Instead of the default of Visual Studio 2015. This will fix the error:

```
Node.js is only supported on Windows 10, Windows Server 2016, or higher.
Setting the NODE_SKIP_PLATFORM_CHECK environment variable to 1 skips this
check, but Node.js might not execute correctly. Any issues encountered on
unsupported platforms will not be fixed.
```